add feature kafka

+++ app.js
@@ -3,9 +3,16 @@ const cors = require('cors');
 const multer = require('multer');
 const path = require('path');
 const fs = require('fs');
-const ocr = require('./utils/ocr');
-const { translate } = require('./utils/translate');
-const { createPDF } = require('./utils/pdf');
+const produceMessage = require('./kafka/config/producer');
+const { resultEmitter } = require('./kafka/config/resultEmitter')
+const { v4: uuidv4 } = require('uuid');
+const { ocrConsumer } = require('./kafka/consumer/consumerOcr');
+const { pdfConsumer } = require('./kafka/consumer/consumerPdf');
+const { translateConsumer } = require('./kafka/consumer/consumerTranslate');
+
+ocrConsumer();
+translateConsumer();
+pdfConsumer();
 
 const app = express();
 const PORT = process.env.PORT || 3000;
@@ -17,7 +24,7 @@ app.use(express.static('public'));
 
 // Cấu hình upload
 const storage = multer.diskStorage({
-  destination: function(req, file, cb) {
+  destination: function (req, file, cb) {
     // Đảm bảo thư mục uploads tồn tại
     const uploadDir = './uploads';
     if (!fs.existsSync(uploadDir)) {
@@ -25,14 +32,14 @@ const storage = multer.diskStorage({
     }
     cb(null, uploadDir);
   },
-  filename: function(req, file, cb) {
+  filename: function (req, file, cb) {
     cb(null, Date.now() + '-' + file.originalname);
   }
 });
 
-const upload = multer({ 
+const upload = multer({
   storage: storage,
-  fileFilter: function(req, file, cb) {
+  fileFilter: function (req, file, cb) {
     // Chỉ chấp nhận file hình ảnh
     if (!file.originalname.match(/\.(jpg|jpeg|png|gif|bmp|tiff)$/)) {
       return cb(new Error('Chỉ chấp nhận file hình ảnh!'), false);
@@ -54,27 +61,32 @@ app.post('/upload', upload.single('image'), async (req, res) => {
     }
 
     const imagePath = req.file.path;
-    
-    // Xử lý OCR: chuyển ảnh sang text
-    const text = await ocr.image2text(imagePath);
-    
-    // Dịch text từ tiếng Anh sang tiếng Việt
-    const viText = await translate(text);
-    
-    // Tạo file PDF từ text tiếng Việt
-    const pdfPath = createPDF(viText);
-    
-    // Trả về file PDF
-    res.download(pdfPath, 'translated_document.pdf', (err) => {
-      if (err) {
-        console.error('Lỗi khi tải file PDF:', err);
-        res.status(500).send('Lỗi khi tải file PDF');
+    const correlationId = uuidv4();
+    // gủi đi 1 topic Kafka
+    await produceMessage('ocr-topic', imagePath, correlationId);
+
+    const TIMEOUT = 15000; // 15 giây
+
+    const timeout = setTimeout(() => {
+      resultEmitter.removeAllListeners(`pdfCreated-${correlationId}`); // tránh leak memory
+      res.status(504).json({ error: 'Xử lý ảnh mất quá nhiều thời gian' });
+    }, TIMEOUT);
+
+    resultEmitter.once(`pdfCreated-${correlationId}`, ({ pdfPath, key }) => {
+      if (key === correlationId) {
+        clearTimeout(timeout);
+        res.download(pdfPath, 'translated_document.pdf', (err) => {
+          if (err) {
+            console.error('Lỗi khi tải file PDF:', err);
+            res.status(500).send('Lỗi khi tải file PDF');
+          }
+
+          // Xóa file ảnh sau khi xử lý xong
+          fs.unlink(imagePath, (err) => {
+            if (err) console.error('Không thể xóa file ảnh tạm:', err);
+          });
+        });
       }
-      
-      // Xóa file ảnh sau khi xử lý xong
-      fs.unlink(imagePath, (err) => {
-        if (err) console.error('Không thể xóa file ảnh tạm:', err);
-      });
     });
   } catch (error) {
     console.error('Lỗi xử lý:', error);
